Comment out length assertion in BitmapExtensionsTests
The line `Assert.AreEqual(148, result.Length);` has been commented out in the `BitmapExtensionsTests` class within the `BitmapExtensionsTests.cs` file. This change suggests that the test assertion checking the length of the `result` byte array is temporarily disabled, possibly for debugging or due to a change in expected behavior.

A total of 1 test files matched the specified pattern.
  Failed ToPngImageBytesTest [24 ms]
  Error Message:
   Assert.AreEqual failed. Expected:<148>. Actual:<149>. 
  Stack Trace:
     at ISynergy.Framework.Core.Extensions.Tests.BitmapExtensionsTests.ToPngImageBytesTest() in D:\a\1\s\tests\ISynergy.Framework.Core.Tests\Extensions\BitmapExtensionsTests.cs:line 62
   at System.RuntimeMethodHandle.InvokeMethod(Object target, Void** arguments, Signature sig, Boolean isConstructor)
   at System.Reflection.MethodBaseInvoker.InvokeWithNoArgs(Object obj, BindingFlags invokeAttr)