Add ContextExtensions with ToEnvironmentalRefreshToken method
Introduced new using directives and namespace `ISynergy.Framework.Core.Extensions`. Created a static class `ContextExtensions` with an extension method `ToEnvironmentalRefreshToken` for the `IContext` interface. This method retrieves and modifies the `RefreshToken` based on the `Environment` property of the `IContext` instance, prefixing it with `GenericConstants.UsernamePrefixTest` for `Test` environment and `GenericConstants.UsernamePrefixLocal` for `Local` environment.