Decouples authentication service
Removes IAuthenticationService from ICommonServices and registers it directly in the dependency injection container, allowing for more flexible and modular use of the authentication service.

The IAuthenticationService is now injected directly into the AuthenticationViewModel and SignInViewModel.

The Application classes no longer subscribe to the AuthenticationChanged and SoftwareEnvironmentChanged events on the authentication service.