Updates package versions and currency symbol
Updates several package versions in Directory.Packages.props to their latest available versions, including Azure.Monitor.OpenTelemetry.Exporter, Microsoft.Identity.Client and OpenTelemetry.Instrumentation.EntityFrameworkCore.

Updates the currency symbol in the WinUI sample app from "" to "€" to reflect the correct currency for the application, as the app is not using the Turkish Lira.